Moved RhinoInside.Revit.GH.AssemblyInfo to Properties/AssemblyInfo.cs
